Week 5 - React Hooks, Fragments
- Move cart state to a custom hook
- Use local storage to persist the cart when the page is refreshed
